Fix: remove incorrect .detach() ops; note true original grads object
The actual original grads are preserved in the grads_list object; grads_act gets messed up because a closure referencing it remains open, and it's non-trivial to close that closure without uninstalling the other hooks in the content manager. Better to just remember this, since I naturally use the constructed grads_list object anyways.